Add files via upload
The user is prompted to enter a number to check for primality.
The program calculates whether the number is prime or not using a simple mathematical check.
The result (either "prime" or "not prime") is displayed to the user.
The user is then asked if they would like to check another number. If they choose "yes," the program continues; otherwise, it exits.